Updated query_bing_search_api.py to create nonexistanf exit folders
query_bing_search_api.py will now check if the provided output path exists and will create a folder for it if it does not. The code was also changed to be more robust so that the user only needs to provide an output name and query_bing_search_api.py will put it into the dataset folder on its own, so that it no longer breaks depending on where it is called from.